Integrated offline imu-monocular SLAM
It likes to crash though, with an issue documented here:
https://github.com/raulmur/ORB_SLAM2/issues/341

I might have to use my own version of orb_slam3 and recompile
everything to fix this, however that would create other issues since for
some reason the version of the library i'm using has a bonus function in
System.cc for retrieving all points in the map.

I would have to put this custom function in the library again after
recompiling I think.