Switching approach for adding scale
Previously, I wanted to compare the change in pose estimated by
orb_slam3 and my apriltag in order to figure out what the scale factor
should be. This proved very unreliable and noisy, so I'm switching
approaches now.

I want to have the user point the camera at the apriltag before
launching the node. As soon as orb_slam3 initiates, the position of the
camera will be at the origin of the world frame, and the position of the
keypoints in the map will be comparable to the known pose of the april
tag. By comparing map points and the apriltag pose at the exact same
time, right when the system is initiated, I should be able to get a
pretty reliable scale factor.